requirements.txt
numpy
pandas
torch
torchvision
pillow
matplotlib
scikit-image